Further improve encapsulation of  time stepper by hiding details in FoamSolver
